+++ src/spec.java
@@ -1,7 +1,7 @@
 
 public class spec {
-	double r,g,b;
-	public spec (double r,double g,double b){
+	float r,g,b;
+	public spec (float r,float g,float b){
 		this.r=r;
 		this.g=g;
 		this.b=b;
@@ -9,16 +9,16 @@ public class spec {
 	public spec mul(spec w){
 		return new spec(w.r*this.r,w.g*this.g,w.b*this.b);
 	}
-	public spec mul(double  w){
+	public spec mul(float  w){
 		return new spec(w*this.r,w*this.g,w*this.b);
 	}
 	public spec add(spec a){
 		return new spec(this.r+a.r,this.g+a.g,this.b+a.b);
 	}
-	public double tol(){
+	public float tol(){
 		return this.r+this.g+this.b;
 	}
-	public void limit(double l){
+	public void limit(float l){
 		if(this.r>l)
 			this.r=l;
 		if(this.r<0)
@@ -33,9 +33,9 @@ public class spec {
 			this.g=0;
 	}
 	public spec sRGB(){
-		double rr=Math.pow(this.r*25,1.0/2.2);
-		double gg=Math.pow(this.g*25,1.0/2.2);
-		double bb=Math.pow(this.b*25,1.0/2.2);
+		float rr=(float)Math.pow(this.r*25f,1.0f/2.2f);
+		float gg=(float)Math.pow(this.g*25f,1.0f/2.2f);
+		float bb=(float)Math.pow(this.b*25f,1.0f/2.2f);
 		return new spec(rr,gg,bb);
 	}
 }
